// src/app/models/userdetails.ts
export interface UserDetail {
  first_name: string;
  email: string;
  full_email: string;
  full_name: string;
  matching_percentage: string;
  skills: string;
  visa_status: string;
  employment_type: string;
  work_type: string;
  preffered_location: string;
  resume_title: string;
  profileId: string;
}